requirements.txt
blinker==1.9.0
click==8.2.1
cs50==9.4.0
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
SQLAlchemy==2.0.41
sqlparse==0.5.3
termcolor==3.1.0
typing_extensions==4.14.0
Werkzeug==3.1.3
wheel==0.45.1
